simplify curl linkage

+++ modules/php.bat
@@ -10,11 +10,7 @@ set INCLUDE=
 set PHP_BUILDTS=0
 
 	REM ~~~~~~~~~~~~ curl
-set CURL_BTYPE=dll
-set CURL_VER=openssl-ssh2_%CURL_BTYPE%_deps-dll
-set PHP_CURL=%PATH_INSTALL%\curl\%CURL_VER%
-if %CURL_BTYPE% == dll (set CURL_LIB_NAME=libcurl.dll)
-if %CURL_BTYPE% == static (set CURL_LIB_NAME=curl.exe)
+set PHP_CURL=%PATH_INSTALL%\curl\openssl
 
 	REM ~~~~~~~~~~~~ php version
 cd /D %PATH_SRC%\php-src
@@ -31,7 +27,7 @@ for %%E in (pecl-memcache pecl-text-xdiff php-ext-brotli xdebug) do (
 	if not exist %PATH_PHP_SDK%\phpmaster\%MSVC_DEPS%\%ARCH%\pecl\%%E\. mklink /J %PATH_PHP_SDK%\phpmaster\%MSVC_DEPS%\%ARCH%\pecl\%%E %PATH_SRC%\%%E
 )
 if not exist %PATH_PHP_SDK%\phpmaster\%MSVC_DEPS%\%ARCH%\build\. mkdir %PATH_PHP_SDK%\phpmaster\%MSVC_DEPS%\%ARCH%\build
-del /Q /F %PATH_PHP_SDK%\phpsdk-local.bat
+if exist %PATH_PHP_SDK%\phpsdk-local.bat del /Q /F %PATH_PHP_SDK%\phpsdk-local.bat
 mklink /H %PATH_PHP_SDK%\phpsdk-local.bat %PATH_MODULES%\phpsdk-local.bat
 
 call %PATH_PHP_SDK%\phpsdk-%MSVC_DEPS%-%ARCH%.bat
